backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic>=2.8.0
pydantic-settings>=2.5.0
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt>=4.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
qrcode==7.4.2
Pillow>=10.2.0
google-generativeai>=0.3.0





